impls/create.rs
This one is difficult since there are several
`impl<T, N> Counter<T, N>` blocks that also create new counters.
