+++ src/impls.rs
@@ -1,5 +1,6 @@
 mod add_iterable;
 mod add_self;
+mod create;
 mod extend;
 mod index;
 mod intersection;
